Extracts common functions to generics
