+++ commands/subsystems/generic.py
@@ -1,16 +1,73 @@
 from telegram.ext import CallbackContext, ConversationHandler
-from telegram import ReplyKeyboardRemove, Update, ReplyKeyboardMarkup
+from telegram import ReplyKeyboardRemove, Update, ReplyKeyboardMarkup, ParseMode
 from .conversation import Conversation
+from ..general import reply_text
 from utils import available_systems, electric_subsystems, mechanics_subsystem
+from spreadsheet import systems
 
 
 # A dictionary to store information about each conversation, identified by the sender's telegram ID
 conversation_task = {}
 
 
+# Project and task listing methods
+# TODO refactor functions
+def get_subtasks(data: list, pos: int, counter: int) -> tuple[str, int, int]:
+    tasks = ""
+    i = pos
+    while i < len(data) and (not data[i][0] or i == pos):
+        if data[i][1] and data[i][2] != "Concluído":
+            tasks += f"{counter} - {data[i][1]}\n"
+            counter += 1
+        i += 1
+    return tasks, i, counter
+
+
+def get_task_lister_text(system: str, subsystem: str) -> str:
+    name = systems[system]["sub"][subsystem]["name"]
+    ss = systems[system]["ss"].sheet(subsystem)
+
+    data = ss.get_all_values()
+    string = f"<b>Subsistema: {name}</b>\n\n<u>Tarefas</u>\n"
+    counter = 1
+
+    for i in range(1, len(data)):
+        if data[i][0]:
+            tasks, pos, counter = get_subtasks(data, i, counter)
+            if tasks:
+                string += f"\n<i>{data[i][0]}</i>\n" + tasks
+            i = pos
+    return string
+
+
 # Returns keyboard markup based on dictionary
 def __create_keyboard(elements: list) -> ReplyKeyboardMarkup:
     return ReplyKeyboardMarkup([elements[i::2] for i in range(2)], one_time_keyboard=True)
+
+
+# Loads configuration and replies text when a system is selected
+def load_system_info(update: Update, selected_system: str) -> any:
+    keyboard = keyboards["subsystem"][selected_system]
+    reply_text(update, f"Sistema {selected_system} selecionado\nInforme o subsistema", keyboard)
+    conversation = get_conversation(update)
+    conversation.system = selected_system
+    conversation.dict = systems[selected_system]["sub"]
+    conversation.ss = systems[selected_system]["ss"]
+
+
+# Loads configuration and replies text when a subsystem is selected
+def load_subsystem_info(update: Update, selected_subsystem: str) -> None:
+    # TODO verify that subsystem is valid
+    conversation = get_conversation(update)
+    conversation.subsystem = selected_subsystem
+    conversation.tasks = get_task_lister_text(conversation.system, selected_subsystem)
+
+    text = (
+        f"<b>Subsistema: {conversation.dict[selected_subsystem]['name']}</b>\n\n"
+        f"{conversation.tasks}\n\n"
+        "Selecione da lista acima o número da tarefa que deseja concluir"
+    )
+    update.message.reply_text(text, reply_markup=ReplyKeyboardRemove(), parse_mode=ParseMode.HTML)
 
 
 # System and subsystem default keyboards
